feature(Loop): add ability to call .fit() multiple times on a model
