+++ src/drift/loop/train.py
@@ -1,8 +1,4 @@
-from __future__ import annotations
-
-from copy import deepcopy
-from enum import Enum
-from typing import Callable, List, Optional, Tuple, Union
+from typing import Callable, List, Optional, Union
 
 import pandas as pd
 from sklearn.base import BaseEstimator
@@ -11,28 +7,14 @@ from ..all_types import TransformationsOverTime
 from ..models.base import Model
 from ..splitters import Split, Splitter
 from ..transformations.base import Composite, Transformation, Transformations
-from ..utils.checks import is_prediction
 from ..utils.list import wrap_in_list
 from .backend.ray import process_transformations as process_transformations_ray
 from .backend.sequential import (
     process_transformations as process_transformations_sequential,
 )
+from .common import deepcopy_transformations, recursively_fit_transform
 from .convenience import replace_transformation_if_not_drift_native
-
-
-class Backend(Enum):
-    no = "no"
-    ray = "ray"
-
-    @staticmethod
-    def from_str(value: Union[str, Backend]) -> Backend:
-        if isinstance(value, Backend):
-            return value
-        for strategy in Backend:
-            if strategy.value == value:
-                return strategy
-        else:
-            raise ValueError(f"Unknown Backend: {value}")
+from .types import Backend, TrainMethod
 
 
 def train(
@@ -43,6 +25,7 @@ def train(
     y: pd.Series,
     splitter: Splitter,
     sample_weights: Optional[pd.Series] = None,
+    train_method: TrainMethod = TrainMethod.parallel,
     backend: Backend = Backend.no,
 ) -> TransformationsOverTime:
     """
@@ -55,26 +38,42 @@ def train(
     )
 
     splits = splitter.splits(length=len(y))
-    process_function = process_transformations_sequential
-    if backend == Backend.ray:
-        process_function = process_transformations_ray
-    processed = process_function(
-        process_transformations_window,
-        transformations,
-        X,
-        y,
-        sample_weights,
-        splits,
-    )
-    idx, only_transformations = zip(*processed)
+    if len(splits) == 0:
+        raise ValueError("No splits were generated by the Splitter.")
+
+    if train_method == TrainMethod.parallel:
+        process_function = process_transformations_sequential
+        if backend == Backend.ray:
+            process_function = process_transformations_ray
+        processed = process_function(
+            process_transformations_window,
+            transformations,
+            X,
+            y,
+            sample_weights,
+            splits,
+        )
+        processed_idx, processed_transformations = zip(*processed)
+    elif train_method == TrainMethod.sequential:
+        only_transformations = [
+            process_transformations_window(
+                X,
+                y,
+                sample_weights,
+                transformations,
+                split,
+            )
+            for split in splits
+        ]
+        idx = [split.model_index for split in splits]
 
     return [
         pd.Series(
             transformation_over_time,
-            index=idx,
+            index=processed_idx,
             name=transformation_over_time[0].name,
         )
-        for transformation_over_time in zip(*only_transformations)
+        for transformation_over_time in zip(*processed_transformations)
     ]
 
 
@@ -101,87 +100,3 @@ def process_transformations_window(
     )
 
     return split.model_index, transformations
-
-
-def deepcopy_transformations(
-    transformation: Union[
-        Transformation, Composite, List[Union[Transformation, Composite]]
-    ]
-) -> Union[Transformation, Composite, List[Union[Transformation, Composite]]]:
-    if isinstance(transformation, List):
-        return [deepcopy_transformations(t) for t in transformation]
-    elif isinstance(transformation, Composite):
-        return transformation.clone(deepcopy_transformations)
-    else:
-        return deepcopy(transformation)
-
-
-def recursively_fit_transform(
-    X: pd.DataFrame,
-    y: pd.Series,
-    sample_weights: Optional[pd.Series],
-    transformations: Transformations,
-) -> pd.DataFrame:
-
-    if isinstance(transformations, List):
-        for transformation in transformations:
-            X = recursively_fit_transform(X, y, sample_weights, transformation)
-        return X
-
-    elif isinstance(transformations, Composite):
-        composite: Composite = transformations
-        # TODO: here we have the potential to parallelize/distribute training of child transformations
-        composite.before_fit(X)
-        results_primary = [
-            recursively_fit_transform(
-                composite.preprocess_X_primary(X, index, y),
-                composite.preprocess_y_primary(y),
-                sample_weights,
-                child_transformation,
-            )
-            for index, child_transformation in enumerate(
-                composite.get_child_transformations_primary()
-            )
-        ]
-
-        if composite.properties.primary_only_single_pipeline:
-            assert len(results_primary) == 1, ValueError(
-                f"Expected single output from primary transformations, got {len(results_primary)} instead."
-            )
-        if composite.properties.primary_requires_predictions:
-            assert is_prediction(results_primary[0]), ValueError(
-                "Expected predictions from primary transformations, but got something else."
-            )
-
-        secondary_transformations = composite.get_child_transformations_secondary()
-        if secondary_transformations is None:
-            return composite.postprocess_result_primary(results_primary)
-        else:
-            results_secondary = [
-                recursively_fit_transform(
-                    composite.preprocess_X_secondary(X, results_primary, index),
-                    composite.preprocess_y_secondary(y, results_primary),
-                    sample_weights,
-                    child_transformation,
-                )
-                for index, child_transformation in enumerate(secondary_transformations)
-            ]
-
-            if composite.properties.secondary_only_single_pipeline:
-                assert len(results_secondary) == 1, ValueError(
-                    f"Expected single output from secondary transformations, got {len(results_secondary)} instead."
-                )
-            if composite.properties.secondary_requires_predictions:
-                assert is_prediction(results_secondary[0]), ValueError(
-                    "Expected predictions from secondary transformations, but got something else."
-                )
-
-            return composite.postprocess_result_secondary(
-                results_primary, results_secondary
-            )
-
-    else:
-        if len(X) == 0:
-            return pd.DataFrame()
-        transformations.fit(X, y, sample_weights)
-        return transformations.transform(X)
