+++ src/drift/models/base.py
@@ -8,12 +8,6 @@ from ..transformations.base import Transformation
 
 class Model(Transformation):
     @abstractmethod
-    def fit(
-        self, X: pd.DataFrame, y: pd.Series, sample_weights: Optional[pd.Series] = None
-    ) -> None:
-        raise NotImplementedError
-
-    @abstractmethod
     def predict(self, X: pd.DataFrame) -> Union[pd.Series, pd.DataFrame]:
         raise NotImplementedError
