+++ src/drift/transformations/test.py
@@ -9,6 +9,7 @@ from .base import Transformation
 class Test(Transformation):
 
     properties = Transformation.Properties(requires_past_X=True)
+    __test__ = False
 
     def __init__(
         self,
